backend/requirements.txt
fastapi==0.115.0
mangum==0.17.0
pydantic==2.8.2
boto3==1.34.162
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pytest==8.3.2
botocore==1.34.162